+++ sdk/messaging/azeventhubs/internal/exported/connection_string_properties.go
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package conn
+
+package exported
 
 import (
 	"errors"
@@ -9,22 +10,34 @@ import (
 	"strings"
 )
 
-type (
-	// ParsedConn is the structure of a parsed Service Bus or Event Hub connection string.
-	ParsedConn struct {
-		Namespace string
-		HubName   string
+// ConnectionStringProperties are the properties of a connection string
+// as returned by [NewConnectionStringProperties].
+type ConnectionStringProperties struct {
+	// Endpoint represents the Endpoint value in the connection string.
+	// Ex: sb://example.servicebus.windows.net
+	Endpoint string
 
-		KeyName string
-		Key     string
+	// EventHubName is EntityPath value in the connection string.
+	EventHubName string
 
-		SAS string
-	}
-)
+	// FullyQualifiedNamespace is the Endpoint value without the protocol scheme.
+	// Ex: example.servicebus.windows.net
+	FullyQualifiedNamespace string
 
-// ParsedConnectionFromStr takes a string connection string from the Azure portal and returns the parsed representation.
-// The method will return an error if the Endpoint, SharedAccessKeyName or SharedAccessKey is empty.
-func ParsedConnectionFromStr(connStr string) (*ParsedConn, error) {
+	// SharedAccessKey is the SharedAccessKey value in the connection string.
+	SharedAccessKey string
+
+	// SharedAccessKeyName is the SharedAccessKeyName value in the connection string.
+	SharedAccessKeyName string
+
+	// SharedAccessSignature is the SharedAccessSignature value in the connection string.
+	SharedAccessSignature string
+}
+
+// NewConnectionStringProperties takes a connection string from the Azure portal and returns the
+// parsed representation The method will return an error if the Endpoint, SharedAccessKeyName
+// or SharedAccessKey is empty.
+func NewConnectionStringProperties(connStr string) (ConnectionStringProperties, error) {
 	const (
 		endpointKey              = "Endpoint"
 		sharedAccessKeyNameKey   = "SharedAccessKeyName"
@@ -37,13 +50,16 @@ func ParsedConnectionFromStr(connStr string) (*ParsedConn, error) {
 	// 1. Connection strings generated from the portal (or elsewhere) that contain an embedded key and keyname.
 	// 2. A specially formatted connection string with an embedded SharedAccessSignature:
 	//   Endpoint=sb://<sb>.servicebus.windows.net;SharedAccessSignature=SharedAccessSignature sr=<sb>.servicebus.windows.net&sig=<base64-sig>&se=<expiry>&skn=<keyname>"
-	var namespace, hubName, keyName, secret, sas string
+	//var namespace, hubName, keyName, secret, sas string
+
+	csp := ConnectionStringProperties{}
+
 	splits := strings.Split(connStr, ";")
 
 	for _, split := range splits {
 		keyAndValue := strings.SplitN(split, "=", 2)
 		if len(keyAndValue) < 2 {
-			return nil, errors.New("failed parsing connection string due to unmatched key value separated by '='")
+			return ConnectionStringProperties{}, errors.New("failed parsing connection string due to unmatched key value separated by '='")
 		}
 
 		// if a key value pair has `=` in the value, recombine them
@@ -53,39 +69,36 @@ func ParsedConnectionFromStr(connStr string) (*ParsedConn, error) {
 		case strings.EqualFold(endpointKey, key):
 			u, err := url.Parse(value)
 			if err != nil {
-				return nil, errors.New("failed parsing connection string due to an incorrectly formatted Endpoint value")
+				return ConnectionStringProperties{}, errors.New("failed parsing connection string due to an incorrectly formatted Endpoint value")
 			}
-			namespace = u.Host
+			csp.Endpoint = value
+			csp.FullyQualifiedNamespace = u.Host
 		case strings.EqualFold(sharedAccessKeyNameKey, key):
-			keyName = value
+			csp.SharedAccessKeyName = value
+			// keyName = value
 		case strings.EqualFold(sharedAccessKeyKey, key):
-			secret = value
+			csp.SharedAccessKey = value
+			// secret = value
 		case strings.EqualFold(entityPathKey, key):
-			hubName = value
+			csp.EventHubName = value
+			//hubName = value
 		case strings.EqualFold(sharedAccessSignatureKey, key):
-			sas = value
+			csp.SharedAccessSignature = value
+			// sas = value
 		}
 	}
 
-	parsed := &ParsedConn{
-		Namespace: namespace,
-		KeyName:   keyName,
-		Key:       secret,
-		HubName:   hubName,
-		SAS:       sas,
+	if csp.FullyQualifiedNamespace == "" {
+		return ConnectionStringProperties{}, fmt.Errorf("key %q must not be empty", endpointKey)
 	}
 
-	if namespace == "" {
-		return parsed, fmt.Errorf("key %q must not be empty", endpointKey)
+	if csp.SharedAccessSignature == "" && csp.SharedAccessKeyName == "" {
+		return ConnectionStringProperties{}, fmt.Errorf("key %q must not be empty", sharedAccessKeyNameKey)
 	}
 
-	if sas == "" && keyName == "" {
-		return parsed, fmt.Errorf("key %q must not be empty", sharedAccessKeyNameKey)
+	if csp.SharedAccessKey == "" && csp.SharedAccessSignature == "" {
+		return ConnectionStringProperties{}, fmt.Errorf("key %q or %q cannot both be empty", sharedAccessKeyKey, sharedAccessSignatureKey)
 	}
 
-	if secret == "" && sas == "" {
-		return parsed, fmt.Errorf("key %q or %q cannot both be empty", sharedAccessKeyKey, sharedAccessSignatureKey)
-	}
-
-	return parsed, nil
+	return csp, nil
 }